# backend/src/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DukaPay API",
    "version": "1.0.0",
    "description": "Backend API for DukaPay lending, scoring, remittance, and indexer flows."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Legacy API base path"
    },
    {
      "url": "/api/v1",
      "description": "Versioned API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Remittance": {
        "type": "object",
        "description": "A single remittance record persisted in the remittances table.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "senderId": {
            "type": "string",
            "example": "GABCDEFGHIJK..."
          },
          "recipientAddress": {
            "type": "string",
            "example": "GABCDEFGHIJK..."
          },
          "amount": {
            "type": "number"
          },
          "fromCurrency": {
            "type": "string",
            "example": "USDC"
          },
          "toCurrency": {
            "type": "string",
            "example": "EURC"
          },
          "memo": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "completed",
              "failed"
            ]
          },
          "transactionHash": {
            "type": "string",
            "nullable": true
          },
          "xdr": {
            "type": "string",
            "description": "Unsigned transaction XDR generated on creation"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "senderId",
          "recipientAddress",
          "amount",
          "fromCurrency",
          "toCurrency",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "RemittancePage": {
        "type": "object",
        "description": "Keyset pagination envelope returned by the remittance list endpoint.",
        "properties": {
          "next_cursor": {
            "type": "string",
            "nullable": true
          },
          "snapshot_seq": {
            "type": "string"
          },
          "total_at_snapshot": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          }
        },
        "required": [
          "next_cursor",
          "snapshot_seq",
          "total_at_snapshot",
          "limit"
        ]
      },
...
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "userId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "loan_approved",
              "repayment_due",
              "repayment_confirmed",
              "loan_defaulted",
              "loan_liquidated",
              "score_changed"
            ]
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "loanId": {
            "type": "integer"
          },
          "actionUrl": {
            "type": "string",
            "nullable": true
          },
          "read": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "type",
          "title",
          "message",
          "read",
          "createdAt"
        ]
      },
      "NotificationsData": {
        "type": "object",
        "properties": {
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            }
          },
          "unreadCount": {
            "type": "integer"
          }
        },
        "required": [
          "notifications",
          "unreadCount"
        ]
      },
      "NotificationsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/NotificationsData"
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
...
    }
  },
  "paths": {
...
    "/notifications": {
      "get": {
        "summary": "Get notifications for the authenticated user",
        "description": "Returns a paginated list of notifications for the authenticated user. Supports filtering by type, status, and date range.\n",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 100
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "loan_approved",
                "repayment_due",
                "repayment_confirmed",
                "loan_defaulted",
                "score_changed"
              ]
            },
            "description": "Filter by notification type"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "unread",
                "read",
                "archived"
              ]
            },
            "description": "Filter by notification status"
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Filter by created_at start date (ISO-8601)"
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Filter by created_at end date (ISO-8601)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of notifications and unread count",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/remittances": {
      "post": {
        "summary": "Create a new remittance",
        "description": "Creates a new remittance record and generates an unsigned XDR transaction for the user to sign with their Stellar wallet (Freighter).\n",
        "tags": [
          "Remittances"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "recipientAddress",
                  "amount",
                  "fromCurrency",
                  "toCurrency"
                ],
                "properties": {
                  "recipientAddress": {
                    "type": "string",
                    "description": "Recipient Stellar public key (56 chars, starts with G)"
                  },
                  "amount": {
                    "type": "number",
                    "description": "Amount to send (in units)"
                  },
                  "fromCurrency": {
                    "type": "string",
                    "enum": [
                      "USDC",
                      "EURC",
                      "PHP"
                    ]
                  },
                  "toCurrency": {
                    "type": "string",
                    "enum": [
                      "USDC",
                      "EURC",
                      "PHP"
                    ]
                  },
                  "memo": {
                    "type": "string",
                    "description": "Optional transaction memo (max 28 chars)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Remittance created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Remittance"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Missing or invalid Bearer token"
          }
        }
      },
      "get": {
        "summary": "Get user's remittances",
        "description": "Returns a paginated list of remittances for the authenticated user. Supports filtering by status, date range, and search by recipient address or reference.\n",
        "tags": [
          "Remittances"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 100
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            },
            "description": "Opaque cursor for pagination"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "processing",
                "completed",
                "failed"
              ]
            },
            "description": "Filter by remittance status"
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Filter by created_at start date (ISO-8601)"
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Filter by created_at end date (ISO-8601)"
          },
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "Search by recipient address or reference"
          }
        ],
        "responses": {
          "200": {
            "description": "Remittances retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Remittance"
                      }
                    },
                    "page": {
                      "$ref": "#/components/schemas/RemittancePage"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "page"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token"
          }
        }
      }
    },
...
  },
...
}